LED Matrix Support
The LED Matrix can now display whatever you want (With colors red, black, and white, for now), by making a 32x8 spreadsheet with each color in the box you want corresponding to the pixel you want, with a comma after it. (Ex. "Re," in (3,3) will display a red pixel in that spot). Scrolling it also works; to disable it, comment out scrollVar.