Added Image Converter
In order to generate an array with the needed layout, create a 32x8 GIF Image (such as with pixilart, resized with GIMP), and then rename it to "image.gif" and place in the Python Image Converter Folder. Then run the image converter in the IDLE shell, and copy and paste the output into the array.